Agregue los modelos de UserAnswer y TextQuestion

+++ source/database/migrations/2015_06_26_212611_create_user_answers_table.php
@@ -15,8 +15,12 @@ class CreateUserAnswersTable extends Migration
         Schema::create('user_answers', function (Blueprint $table) {
             $table->increments('id');
             $table->string('answer');
+            // FK to survey_respondents
             $table->integer('survey_respondent_id')->unsigned();
             $table->foreign('survey_respondent_id')->references('id')->on("survey_respondents");
+            // FK to survey
+            $table->integer('survey_id')->unsigned();
+            $table->foreign('survey_id')->references('id')->on("surveys");
             $table->timestamps();
         });
     }
@@ -28,6 +32,9 @@ class CreateUserAnswersTable extends Migration
      */
     public function down()
     {
+      if(Schema::hasTable('user_answers'))
+      {
         Schema::drop('user_answers');
+      }
     }
 }
